src/test/java/com/codingquestions/app/BitRepresentationAndOperation/ReverseBitsOf32BitIntegerTest.java
package com.codingquestions.app.BitRepresentationAndOperation;

import static org.junit.jupiter.api.Assertions.assertEquals;
import static org.junit.jupiter.api.Assertions.assertTrue;

import org.junit.jupiter.api.AfterAll;
import org.junit.jupiter.api.AfterEach;
import org.junit.jupiter.api.Assertions;
import org.junit.jupiter.api.BeforeAll;
import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Disabled;
import org.junit.jupiter.api.Tag;
import org.junit.jupiter.api.Test;

// import the class that you want to test here
import com.codingquestions.app.BitRepresentationAndOperation.ReverseBitsOf32BitInteger;

public class ReverseBitsOf32BitIntegerTest {

    @BeforeAll
    static void setup() {
    }

    @BeforeEach
    void setupThis() {
    }

    @Test
    void testOne() {
        ReverseBitsOf32BitInteger instance = new ReverseBitsOf32BitInteger();
        assertEquals(0L, instance.reverseBits(0L));
    }

    @AfterEach
    void tearThis() {
    }

    @AfterAll
    static void tear() {
    }
}